requirements.txt
faiss-cpu
numpy
openai
pypdf
python-dotenv
sentence-transformers
